Advancing to v0.3.3 (enabling compilation on Android)
First modifications to enable the compilation on Android systems.
'udphdr' is defined as '__kernel_udphdr' as the latter may be the name
of the UDP header structure in 'linux/udp.h' which is available on
Android systems.
Moreover, all the calls to 'bzero' have been replaced with the standard
C function 'memset'.
Tested on Termux with Android 8 and 9.